% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_ocean_data.R
\name{animate_ocean_data}
\alias{animate_ocean_data}
\title{Animate ocean data}
\usage{
animate_ocean_data(input, color = "SST", ...)
}
\arguments{
\item{input}{input}

\item{color}{variable for color}

\item{...}{passes functions}
}
\description{
DEVELOPMENT function to animate ocean data maps
}
