% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ocean_data.R
\name{check_ocean_data}
\alias{check_ocean_data}
\title{Check ocean data}
\usage{
check_ocean_data(input, ...)
}
\arguments{
\item{input}{input}

\item{...}{passes functions}
}
\description{
DEVELOPMENT function to distill check data
}
