% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_ocean_data.R
\name{distill_ocean_data}
\alias{distill_ocean_data}
\title{Distill ocean data}
\usage{
distill_ocean_data(input, var, group = "all", ...)
}
\arguments{
\item{input}{input}

\item{var}{variable}

\item{group}{group factor}

\item{...}{passes functions}
}
\description{
DEVELOPMENT function to distill ocean data maps
}
