% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOD.R
\name{download_od}
\alias{download_od}
\title{download_OD
function to download data}
\usage{
download_od(data_id, data_var, space, time, ...)
}
\arguments{
\item{data_id}{id}

\item{data_var}{var}

\item{space}{id}

\item{time}{var}
}
\description{
download_OD
function to download data
}
