% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ocean_data.R
\name{export_ocean_data}
\alias{export_ocean_data}
\title{Export ocean data}
\usage{
export_ocean_data(input, output, ...)
}
\arguments{
\item{input}{input}

\item{output}{file format for output}

\item{...}{passes functions}
}
\description{
DEVELOPMENT function to export ocean data
}
