% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_functions.R
\name{extract_griddap_180_points}
\alias{extract_griddap_180_points}
\title{extract griddap 180_points}
\usage{
extract_griddap_180_points(
  space = NULL,
  time = NULL,
  this_dataset = NULL,
  this_var = NULL
)
}
\arguments{
\item{space}{space}

\item{time}{time}

\item{this_dataset}{= this dataset}

\item{this_var}{\itemize{
\item this var
}}
}
\description{
extract griddap 180_points
}
