% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOD.R
\name{get_xml_headers}
\alias{get_xml_headers}
\title{get XML header}
\usage{
get_xml_headers(
  data_id,
  data_var,
  burl = "https://coastwatch.pfeg.noaa.gov/erddap/",
  space,
  time
)
}
\arguments{
\item{data_id}{1:52 common datasets}

\item{burl}{base URL  "https://coastwatch.pfeg.noaa.gov/erddap/",}

\item{space}{space}

\item{time}{time}
}
\description{
get XML header
}
