% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_ocean_data.R
\name{interactive_ocean_data}
\alias{interactive_ocean_data}
\title{Map ocean data}
\usage{
interactive_ocean_data(input, color = "SST", rows = 1, ...)
}
\arguments{
\item{input}{input}

\item{color}{variable for map color}

\item{rows}{number of rows if faceted}

\item{...}{passes functions}
}
\description{
DEVELOPMENT function to generate ocean data maps
}
