% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ocean_data.R
\name{map_ocean_data}
\alias{map_ocean_data}
\title{Map ocean data}
\usage{
map_ocean_data(input, color = "SST", rows = 1, ...)
}
\arguments{
\item{input}{input}

\item{color}{variable for color}

\item{rows}{number of rows if faceted}

\item{...}{passes functions}
}
\description{
DEVELOPMENT function to generate ocean data maps
}
