% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_functions.R
\name{myURL.4dim}
\alias{myURL.4dim}
\title{myURL.4dim}
\usage{
myURL.4dim(
  burl = "https://coastwatch.pfeg.noaa.gov/erddap/",
  data_id,
  ftype = ".nc",
  data_var,
  space,
  time
)
}
\arguments{
\item{burl}{base URL  "https://coastwatch.pfeg.noaa.gov/erddap/",}

\item{data_id}{data id}

\item{ftype}{file type = ".nc",}

\item{data_var}{data var,}

\item{space}{space}

\item{time}{time}

\item{dataset}{1:52 common datasets}
}
\description{
get 4 dim URL (lon/lat/z/var)
}
