% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOD.R
\name{tidy_od}
\alias{tidy_od}
\title{tidy_OD
function to tidy output from erdapp}
\usage{
tidy_od(file.data, data_id, data_var, sites = FALSE, info = info.erddap, ...)
}
\arguments{
\item{file.data}{file}

\item{data_id}{id}

\item{data_var}{var}

\item{sites}{sites or coords? TRUE=sites, FALSE=coords}
}
\description{
tidy_OD
function to tidy output from erdapp
}
